# openapi.yaml
openapi: 3.1.0
info:
  title: Veilence-MX API
  description: >
    Enterprise SaaS supply chain compromise monitor for PyPI and npm packages.
    Provides endpoints for package monitoring, release analysis, alert management,
    organization/RBAC, notifications, and system administration.
  version: 1.0.0
  contact:
    name: Veilence-MX Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.veilence.example.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/invitations/{token}:
    get:
      tags: [Organizations]
      summary: Get invitation info (public)
      description: >
        Returns invitation details so the frontend can display "You've been invited
        to {org}" before the user logs in. No authentication required.
      security: []
      operationId: getInvitationInfo
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation info
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                      orgId:
                        type: integer
                      expiresAt:
                        type: string
                        format: date-time
                      accepted:
                        type: boolean
                      expired:
                        type: boolean
                  error:
                    type: string
                    nullable: true
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Dashboard (org-scoped) ─────────────────────────────────────
...
  /api/orgs/{orgId}/members:
    get:
      tags: [Organizations]
      summary: List members
      description: Lists organization members with roles and user info. Requires `members:read` permission.
      operationId: listMembers
      parameters:
        - $ref: "#/components/parameters/OrgIdPath"
      responses:
        "200":
          description: List of members (flattened with email, firstName, lastName at top level)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_MemberList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/orgs/{orgId}/members/{userId}:
    delete:
      tags: [Organizations]
      summary: Remove member
      description: Removes a member from the organization. Requires `members:remove` permission.
      operationId: removeMember
      parameters:
        - $ref: "#/components/parameters/OrgIdPath"
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/orgs/{orgId}/members/{userId}/role:
    put:
      tags: [Organizations]
      summary: Update member role
      description: Changes a member's role. Requires `members:remove` permission.
      operationId: updateMemberRole
      parameters:
        - $ref: "#/components/parameters/OrgIdPath"
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleId]
              properties:
                roleId:
                  type: integer
      responses:
        "200":
          description: Member role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_OrgMember"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/orgs/{orgId}/invitations:
    post:
      tags: [Organizations]
      summary: Invite member
      description: Invites a user by email to the organization. Requires `members:invite` permission.
      operationId: inviteMember
      parameters:
        - $ref: "#/components/parameters/OrgIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Invitation created (includes token for URL construction)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Invitation"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [Organizations]
      summary: List pending invitations
      description: Lists pending invitations for the organization. Requires `members:read` permission.
      operationId: listPendingInvitations
      parameters:
        - $ref: "#/components/parameters/OrgIdPath"
      responses:
        "200":
          description: List of pending invitations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_InvitationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/orgs/{orgId}/invitations/{id}:
    delete:
      tags: [Organizations]
      summary: Revoke invitation
      description: Revokes a pending invitation. Requires `members:invite` permission.
      operationId: revokeInvitation
      parameters:
        - $ref: "#/components/parameters/OrgIdPath"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Invitation revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Message"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/orgs/{orgId}/invitations/{token}/accept:
    post:
      tags: [Organizations]
      summary: Accept invitation
      description: >
        Accepts an invitation to join an organization. Requires authentication
        but not org membership. The accepting user's email must match the invitation.
      operationId: acceptInvitation
      parameters:
        - $ref: "#/components/parameters/OrgIdPath"
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation accepted - user is now a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_OrgMember"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          description: Invitation expired
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login or /api/auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key created via /api/auth/api-keys
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
    OrgIdPath:
      name: orgId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        data:
          nullable: true
        error:
          type: string

    # ── Domain Schemas ──────────────────────────────────────────────
...
    OrgMember:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        userId:
          type: integer
        roleId:
          type: integer
        role:
          $ref: "#/components/schemas/Role"
        joinedAt:
          type: string
          format: date-time
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    Role:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        name:
          type: string
        description:
          type: string
        isSystem:
          type: boolean
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
    Permission:
      type: object
      properties:
        id:
          type: integer
        resource:
          type: string
        action:
          type: string
...
    Invitation:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        email:
          type: string
        roleId:
          type: integer
        invitedBy:
          type: integer
        expiresAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    InviteMemberRequest:
      type: object
      required: [email, roleId]
      properties:
        email:
          type: string
          format: email
        roleId:
          type: integer
...
    ApiResponse_Message:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
        error:
          type: string
          nullable: true
...
    ApiResponse_OrgMember:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/OrgMember"
        error:
          type: string
          nullable: true
    ApiResponse_MemberList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OrgMember"
        error:
          type: string
          nullable: true
...
    ApiResponse_Invitation:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: integer
            orgId:
              type: integer
            email:
              type: string
            roleId:
              type: integer
            token:
              type: string
              description: Raw invitation token for URL construction
            invitedBy:
              type: integer
            expiresAt:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
        error:
          type: string
          nullable: true
    ApiResponse_InvitationList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invitation"
        error:
          type: string
          nullable: true
...